requirements.txt
requests
beautifulsoup4
lxml
feedparser
PyYAML
feedparser
beautifulsoup4
python-dateutil
requests
